+++ src/article/title for your post.md
@@ -1,0 +1,8 @@
+---
+title: "title for your post"
+date: "2022-09-29"
+category: "TIL"
+imgPath: "../images/gatsby-icon.png"
+---
+
+**FROM HERE, WRITE CONTENTS USING MARKDOWN SYNTAX
