+++ src/components/post/downloadTxtFile.js
@@ -1,0 +1,23 @@
+const downloadTxtFile = () => {
+  const element = document.createElement("a");
+  let data = document.querySelector(".w-md-editor-text-input").value
+  let work1 = data.split('\n');
+  let title = work1.filter(el => el.match(/title:/)).toString().substring(7).substring(1).slice(0, -1);
+  const file = new Blob(
+    [data],
+    {
+      type: "text/plain",
+    }
+  );
+  // let reader = new FileReader();
+  // reader.onload = function() {
+  //   alert(reader.result);
+  // }
+  // reader.readAsText(blob)
+
+  element.href = URL.createObjectURL(file);
+  element.download = title + '.md';
+  document.body.appendChild(element);
+  element.click();
+};
+export default downloadTxtFile;
